pyproject.toml
[project]
name = "testclaw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chuk-mcp>=0.9.1",
    "eventlet>=0.40.4",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-socketio>=5.6.1",
    "glib>=1.0.0",
    "httpx>=0.28.1",
    "markdown>=3.10.2",
    "matplotlib>=3.10.8",
    "mcp-client>=0.0.0",
    "mcp-open-client>=0.9.14",
    "modelcontextprotocol>=1.0.1",
    "numpy>=2.4.3",
    "openai>=2.28.0",
    "pillow>=12.1.1",
    "py-mcp-client>=0.0.8",
    "pyautogui>=0.9.54",
    "pyreadline3>=3.5.4",
    "pywayland>=0.4.18",
    "requests>=2.33.0",
    "termcolor>=3.3.0",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
    "urllib3==1.26.18",
]
